requirements.txt
streamlit
requests
gTTS
pyngrok
git+https://github.com/openai/whisper.git
sievedata
ffmpeg-python



